requirements.txt
azure-storage-blob==12.5.0
detecto==1.1.6
matplotlib==3.2.1
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.2
pathlib==1.0.1
Pilo==0.6.2
seaborn==0.10.1
shapely==1.7.1
torch==1.6.0
torchvision==0.7.0
xmltodict==0.12.0